mutações alteradas para atender o crossomo do TOP

+++ solution/FunctionObjective.py
@@ -58,7 +58,7 @@ class FunctionObjective:
         return ((prizes_total ** 3) - coust_total) * -1
 
     def coust_insert(self, chromosome):
-        prizes_total = self.prizes.take(chromosome).sum()
+        prizes_total = self.prizes.take(chromosome.astype(int)).sum()
         coust_total = self.med_custo(chromosome)
         return ((prizes_total ** 2)/ coust_total)
